+++ src/core/hle/kernel/shared_page.cpp
@@ -50,8 +50,7 @@ static std::chrono::seconds GetInitTime() {
     case Settings::InitClock::FixedTime:
         return std::chrono::seconds(Settings::values.init_time);
     default:
-        UNREACHABLE_MSG("Invalid InitClock value ({})",
-                        static_cast<u32>(Settings::values.init_clock));
+        UNREACHABLE_MSG("Invalid InitClock value ({})", Settings::values.init_clock);
     }
 }
